layered pattern method별 -> 기능별 묶음으로 수정
