+++ teamB/jungJaeUk/app.js
@@ -2,9 +2,7 @@ const http = require('http');
 const express = require('express');
 const cors = require('cors');
 const morgan = require('morgan');
-
-const dotenv = require('dotenv');
-dotenv.config();
+require('dotenv').config();
 
 const routes = require("./routes");
