--- teamB/jungJaeUk/controllers/userConSignup.js
@@ -1,20 +1,0 @@
-const userService = require("../services/userSvcSignup");
-
-const signUp = async (req, res) => {
-  try {
-    const { name, email, profile_image, password } = req.body;
-
-    if ( !name || !email || !profile_image || !password ) {
-      return res.status(400).json({ "message" : "KEY_ERROR" });
-    }
-
-    await userService.signUp( name, email, profile_image, password );
-
-    res.status(201).json({ "data" : "userCreated" });
-  } catch (err) {
-    console.log(err);
-    return res.status(err.statusCode || 500).json({ "message" : err.message });
-  }
-};
-
-module.exports = { signUp };
