+++ teamB/jungJaeUk/routes/userRoute.js
@@ -1,9 +1,8 @@
 const router = require("express").Router();
 
-const userConSignup = require("../controllers/userConSignup");
-const userConPosts = require("../controllers/userConPosts");
+const userController = require("../controllers/userController");
 
-router.post("/signup", userConSignup.signUp);
-router.get("/:id/posts", userConPosts.posts);
+router.post("/signup", userController.signUp);
+router.get("/:id/posts", userController.posts);
 
 module.exports = { router };
