--- teamB/jungJaeUk/services/postSvcAll.js
@@ -1,9 +1,0 @@
-const userDao = require("../models/postDaoAll");
-
-const postsAll = async () => {
-  const postsAll = await userDao.postsAll();
-
-  return postsAll;
-};
-
-module.exports = { postsAll };
