--- teamB/jungJaeUk/services/userSvcPosts.js
@@ -1,9 +1,0 @@
-const userDao = require("../models/userDaoPosts");
-
-const posts = async ( userId ) => {
-  const postsList = await userDao.postsList(userId);
-
-  return postsList;
-};
-
-module.exports = { posts };
